Fixed bug where files larger than 2147483647 bytes (2147.48 MB) crashed program as they were larger than Int32 max value.

+++ ApiClient/RestApiResponses/FolderContentsResponse.cs
@@ -121,7 +121,7 @@ public class FolderContentsResponseIncludedAttributes
 
     [JsonProperty("versionNumber")] public int VersionNumber { get; set; }
 
-    [JsonProperty("storageSize")] public int StorageSize { get; set; }
+    [JsonProperty("storageSize")] public long StorageSize { get; set; }
 
     [JsonProperty("fileType")] public string FileType { get; set; }
